Move unauthenticated endpoints to `backend/app/api.py` and add `authenticate_user` dependency.
Unauthenticated users are redirected to login page (full reload of SPA occurs). Sessions return a null value when session_id is not present in redis.
Update client-side API calls and backend tests to handle url changes.
