+++ backend/app/api/endpoints/users.py
@@ -1,12 +1,7 @@
-from urllib.parse import urljoin
-
-import aiohttp
-from fastapi import APIRouter, HTTPException, Request
+from fastapi import APIRouter, Request
 from starlette.responses import JSONResponse, RedirectResponse
 
 from app.api.deps import fastapi_sessions
-from app.api.models import OpenIdInfo
-from app.core.http_client import http_client
 from app.core.models import ServerSessionData
 
 router = APIRouter()
@@ -15,31 +10,6 @@ router = APIRouter()
 @router.get("/")
 async def message_users():
     return {"message": "Hello User!!!"}
-
-
-@router.post("/verify-openid")
-async def verify_openid(request: Request, open_id_info: OpenIdInfo):
-    matrix_homeserver = "https://" + open_id_info.matrix_server_name
-    params = {"access_token": open_id_info.access_token}
-    url = urljoin(matrix_homeserver, "/_matrix/federation/v1/openid/userinfo")
-
-    try:
-        async with http_client.session.get(url, params=params) as resp:
-            if resp.status != 200:
-                raise HTTPException(status_code=404, detail="Invalid token")
-            data = await resp.json()
-
-            server_session_data = ServerSessionData(
-                matrix_user=data["sub"], matrix_homeserver=matrix_homeserver
-            )
-
-            response = JSONResponse({"message": "success"})
-            # Creating a server session with the matrix username.
-            response = fastapi_sessions.create_session(response, data=server_session_data)
-            return response
-
-    except aiohttp.ClientConnectionError as err:
-        raise HTTPException(status_code=500, detail="Internal server error")
 
 
 @router.post("/logout")
@@ -67,5 +37,4 @@ async def change_tokens(request: Request):
 # Redirects to the React Home page
 @router.get("/redirect")
 async def redirect():
-    response = RedirectResponse(url="/")
-    return response
+    return RedirectResponse("/")
